Rename project to appdir
We use `appdir` everywhere else, and the prefix was getting annoying.
It's likely we will get some confusion with `appdirs`, but given that
that is written in python (and has numerous clones in other languages
names the same thing), we shouldn't have too much of a problem. We also
don't have users yet, so no harm done.