Support both building archives and installing systemwide
The install() structure for creating an archive and installing is
completely different. Introduce a variable (CREATE_ARCHIVE) to keep
track of what we're building for. I would have liked to support building
for both without reconfiguring (considering it's the same files just
going to a different place), but cmake doesn't seem to support this
without significant wrangling.